requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
blinker==1.9.0
click==8.1.8
discord.py==2.5.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
Werkzeug==3.1.3
yarl==1.18.3
PS E:\Botdsdsdsdsdsd> pip freeze requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
discord.py==2.5.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
Werkzeug==3.1.3
yarl==1.18.3